+++ CHANGELOG.md
@@ -15,3 +15,8 @@ First release.
 - Templates: MacVLAN launch-xrd example, xr-compose template
 - Sample xr-compose topologies: 'simple-bgp', 'bgp-ospf-triangle, 'segment-routing'
 - Tests: host-check UT
+
+
+### v1.0.1 (2022-09-06)
+
+Emit a warning in `host-check` when 2M hugepages are used.